login to vue + permission

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Api\AuthController;
 use App\Http\Controllers\Api\DataController;
 use App\Http\Controllers\Api\UserController;
 use Illuminate\Http\Request;
@@ -20,12 +21,14 @@ Route::prefix('data')->group(function() {
     Route::get('/roles', [DataController::class, 'roles']);
 });
 
+Route::post('/login', [AuthController::class, 'login']);
+
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
 Route::prefix('users')->middleware('auth:sanctum')->group(function() {
-    Route::get('/', [UserController::class, 'index'])->middleware(['role:super_admin,admin']);
-    Route::post('/', [UserController::class, 'store'])->middleware(['role:super_admin']);
-    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware(['role:super_admin']);
+    Route::get('/', [UserController::class, 'index'])->middleware(['role:super_admin,admin,user']);
+    Route::post('/', [UserController::class, 'store'])->middleware(['role:super_admin,admin']);
+    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware(['role:super_admin,admin']);
 });
